requirements.txt
groq
python-dotenv
google-genai
flask